main/index.tsx 모듈 분리
